Ce composant définit à chaque render faisait sauter l'input
Oups, j'aurais peut-être juste du faire un mémo !
